import error and trigger dag
Import error api polling 30 sec
delete for import error
changed name to schedule error
trigger the dag disable on pause

+++ dataproc_jupyter_plugin/services/dagListService.py
@@ -64,7 +64,7 @@ class DagListService():
     
 
 class DagDeleteService():
-    def delete_job(self, credentials, composer_name, dag_id,log):
+    def delete_job(self, credentials, composer_name, dag_id,from_page,log):
         airflow_uri, bucket = DagListService.get_airflow_uri(self,composer_name,credentials,log)
         if 'access_token' and 'project_id' and 'region_id' in credentials:
             access_token = credentials['access_token']
@@ -75,8 +75,9 @@ class DagDeleteService():
             'Content-Type': CONTENT_TYPE,
             'Authorization': f'Bearer {access_token}'
             }
-            response = requests.delete(api_endpoint,headers=headers)
-            log.info(response)
+            if from_page != "importErrorPage":
+                response = requests.delete(api_endpoint,headers=headers)
+                log.info(response)
             cmd = f"gsutil rm gs://{bucket}/dags/dag_{dag_id}.py"
             process = subprocess.Popen(cmd, stdout=subprocess.PIPE, stderr=subprocess.PIPE, shell=True)
             output, _ = process.communicate()
